refactor: rss 수집기 연쇄 책임 패턴을 제거한다
