--- app-batch/app-batch-rss/src/main/kotlin/site/techmoa/app/batch/rss/collector/rssreader/RssReaderCollector.kt
@@ -1,21 +1,0 @@
-package site.techmoa.app.batch.rss.collector.rssreader
-
-import org.springframework.stereotype.Component
-import site.techmoa.app.batch.rss.collector.RssCollector
-import site.techmoa.app.batch.rss.collector.rssreader.handler.RssCollectContext
-import site.techmoa.app.batch.rss.collector.rssreader.handler.RssCollectHandler
-import site.techmoa.app.batch.rss.collector.rssreader.handler.RssCollectHandlerChain
-
-@Component
-class RssReaderCollector(
-    handlers: List<RssCollectHandler>,
-) : RssCollector {
-
-    private val orderedHandlers = handlers.sortedBy { it.getOrder() }
-
-    override fun execute() {
-        val context = RssCollectContext()
-        val chain = RssCollectHandlerChain(orderedHandlers)
-        chain.next(context)
-    }
-}
