--- app-batch/app-batch-rss/src/main/kotlin/site/techmoa/app/batch/rss/collector/rssreader/handler/RssCollectContext.kt
@@ -1,11 +1,0 @@
-package site.techmoa.app.batch.rss.collector.rssreader.handler
-
-import com.apptasticsoftware.rssreader.Item
-import site.techmoa.app.storage.db.entity.ArticleEntity
-import site.techmoa.app.storage.db.entity.BlogEntity
-
-data class RssCollectContext(
-    var blogs: List<BlogEntity> = emptyList(),
-    var collectedItems: Map<BlogEntity, List<Item>> = emptyMap(),
-    var newArticles: List<ArticleEntity> = emptyList(),
-)
