--- app-batch/app-batch-rss/src/main/kotlin/site/techmoa/app/batch/rss/collector/rssreader/handler/RssCollectHandler.kt
@@ -1,6 +1,0 @@
-package site.techmoa.app.batch.rss.collector.rssreader.handler
-
-interface RssCollectHandler {
-    fun getOrder(): Int
-    fun handle(context: RssCollectContext, chain: RssCollectHandlerChain)
-}
