+++ app-batch/app-batch-rss/src/main/kotlin/site/techmoa/app/batch/rss/scheduler/LocalCollectorScheduler.kt
@@ -3,7 +3,7 @@ package site.techmoa.app.batch.rss.scheduler
 import org.springframework.context.annotation.Profile
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
-import site.techmoa.app.batch.rss.collector.RssCollector
+import site.techmoa.app.batch.rss.RssCollector
 
 @Component
 @Profile("default")
